pyproject.toml
[project]
name = "aco-learner"
version = "0.1.0"
description = "Convert documents (PDF, etc.) to voice-agent-ready scripts and audio"
requires-python = ">=3.10"
dependencies = [
    "pymupdf4llm>=0.0.10",
    "litellm>=1.40.0",
    "streamlit>=1.35.0",
    "edge-tts>=6.1.0",
    "pydub>=0.25.1",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
kokoro = ["kokoro>=0.3.0", "soundfile>=0.13.0"]
openai-tts = ["openai>=1.30.0"]
dev = ["pytest>=8.0.0", "pytest-asyncio>=0.23.0"]
all = ["kokoro>=0.3.0", "openai>=1.30.0"]

[project.scripts]
aco-learner = "app.ui.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
